HomeView is now its own component
